extract the cache population  into cron, cache whole filesystem to improve performance
