+++ src/files/files.controller.ts
@@ -1,16 +1,28 @@
-import { Controller, Get } from '@nestjs/common';
+import { CACHE_MANAGER } from '@nestjs/cache-manager';
+import { Controller, Get, Inject } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
-import { Observable } from 'rxjs';
+import { Cache } from 'cache-manager';
+import { Observable, from, of, switchMap } from 'rxjs';
 import { DirectoryEntity } from './entity';
 import { FilesService } from './files.service';
+import { Directory } from './interfaces';
 
 @ApiTags('Files')
 @Controller('files')
 export class FilesController {
-  constructor(private readonly filesService: FilesService) {}
+  constructor(
+    private readonly filesService: FilesService,
+    @Inject(CACHE_MANAGER) private readonly cacheManager: Cache,
+  ) {}
 
   @Get()
   public findAll(): Observable<DirectoryEntity> {
-    return this.filesService.findAll();
+    return from(this.cacheManager.get<Directory>('fileSystem')).pipe(
+      switchMap((val) => {
+        if (val) return of(val);
+
+        return this.filesService.findAll();
+      }),
+    );
   }
 }
